tempo-test/requirements.txt
flask==2.0.1
werkzeug==2.0.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0 